Start fixing TS-480 backend. Make TS-890S backend separate from TS-480.
