+++ rigs/kenwood/Makefile.am
@@ -1,7 +1,7 @@
 
 TSSRC = ts850.c ts870s.c ts570.c ts450s.c ts950.c ts50s.c \
 	ts790.c ts2000.c k2.c k3.c xg3.c ts930.c \
-	ts680.c ts690.c ts140.c ts480.c trc80.c ts590.c \
+	ts680.c ts690.c ts140.c ts480.c trc80.c ts590.c ts890s.c \
 	ts990s.c ts990s.h flex6xxx.c pihpsdr.c
 
 IC10SRC = ts440.c ts940.c ts711.c ts811.c r5000.c
